Refactor GitHub client extensions
