+++ lib/github.js
@@ -1,5 +1,8 @@
-const Bottleneck = require('bottleneck')
 const Octokit = require('@octokit/rest')
+
+const addPagination = require('./github/pagination')
+const addRateLimiting = require('./github/pagination')
+const addLogging = require('./github/pagination')
 
 /**
  * the [@octokit/rest Node.js module](https://github.com/octokit/rest.js),
@@ -9,43 +12,12 @@ const Octokit = require('@octokit/rest')
  * @typedef github
  * @see {@link https://github.com/octokit/rest.js}
  */
-
-const defaultCallback = response => response
-async function paginate (octokit, responsePromise, callback = defaultCallback) {
-  let collection = []
-  let getNextPage = true
-  let done = () => {
-    getNextPage = false
-  }
-  let response = await responsePromise
-  collection = collection.concat(await callback(response, done))
-  // eslint-disable-next-line no-unmodified-loop-condition
-  while (getNextPage && octokit.hasNextPage(response)) {
-    response = await octokit.getNextPage(response)
-    collection = collection.concat(await callback(response, done))
-  }
-  return collection
-}
-
 function EnhancedGitHubClient (options) {
   const octokit = Octokit(options)
-  const limiter = options.limiter || new Bottleneck({ maxConcurrent: 1, minTime: 1000 })
-  const logger = options.logger
-  const noop = () => Promise.resolve()
 
-  octokit.hook.before('request', limiter.schedule.bind(limiter, noop))
-  octokit.hook.error('request', (error, options) => {
-    const {method, url, headers, ...params} = options
-    const msg = `GitHub request: ${method} ${url} - ${error.code} ${error.status}`
-    logger.debug({params}, msg)
-    throw error
-  })
-  octokit.hook.after('request', (result, options) => {
-    const {method, url, headers, ...params} = options
-    const msg = `GitHub request: ${method} ${url} - ${result.meta.status}`
-    logger.debug({params}, msg)
-  })
-  octokit.paginate = paginate.bind(null, octokit)
+  addRateLimiting(octokit, options.limiter)
+  addLogging(octokit, options.logger)
+  addPagination(octokit)
 
   return octokit
 }
